TMP: refactor dram class

+++ memory/demo/memory_demo.cc
@@ -4,7 +4,7 @@
 int main() {
   std::array<xr5::types::Byte, 64> p = {56};
   p.fill(5);
-  for (int i = 0; i < p.size(); i++)
+  for (size_t i = 0; i < p.size(); i++)
     std::cout << static_cast<int>(p[i]) << "\n";
   return 0;
 }
